safari-600.7-branch/LayoutTests/js/typed-array-copy-expected.txt

Reduced test case for https://bugs.webkit.org/show_bug.cgi?id=83818

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a[0] is 0x0505
PASS a[0] is 0x0005
PASS a[1] is 0x0005
PASS successfullyParsed is true

TEST COMPLETE
